+++ framework/bunrun/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @forklaunch/bunrun
 
+## 0.0.15
+
+### Patch Changes
+
+- Update package versions, and add x-powered-by forklaunch
+
+## 0.0.14
+
+### Patch Changes
+
+- update internal package versions
+
+## 0.0.13
+
+### Patch Changes
+
+- update package versions
+
 ## 0.0.12
 
 ### Patch Changes